--- pages/register/register2copy.js
@@ -1,117 +1,0 @@
-import Head from 'next/head'
-import Image from 'next/image'
-import { getApps, getApp, initializeApp } from "firebase/app";
-import { getAuth, createUserWithEmailAndPassword, updateProfile } from "firebase/auth";
-import styles from "/styles/Home.module.css";
-import {firebaseConfig} from '../../settings/firebaseConfig';
-import myImage from '../../public/pic/welcome.png';
-import React, { useEffect, useState } from 'react';
-
-
-
-
-if (getApps().length === 0) {
-  initializeApp(firebaseConfig);
-}
-const Home = () => {
-  const [account, setAccount] = useState({ email: "", password: "", displayName: "" });
-  const [message, setMessage] = useState("");
-  const auth = getAuth();  
-  const handleChange = function (e) {
-    setAccount({ ...account, [e.target.name]: e.target.value })
-  }
-  async function haddleOnClick(){
-  const res = await createUserWithEmailAndPassword(auth, account.email, account.password);
-  updateProfile(auth.currentUser, {
-    displayName: "Jane Q. User", photoURL: "https://example.com/jane-q-user/profile.jpg"
-  }).then(() => {
-    // Profile updated!
-    // ...
-  }).catch((error) => {
-    // An error occurred
-    // ...
-  });
-
-  try {
-    const res = await createUserWithEmailAndPassword(
-      auth,
-      account.email,
-      account.password
-    );
-    setMessage("帳號已產生");
-    console.log({ res });
-  } catch (error) {
-    let message = "";
-    switch (error.code) {
-      case "auth/email-already-in-use":
-        message = "電子信箱已註冊";
-        break;
-      case "auth/weak-password":
-        message = "密碼強度不足";
-        break;
-      case "auth/invalid-email":
-        message = "電子郵件格式錯誤";
-        break;
-      default:
-        message = "系統錯誤:" + error.code;
-    }
-    setMessage(message);
-  }
-}
-
-  return (
-    <div className={styles.container}>
-      <Head>
-      <title>筆記分享區</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <main className={styles.main}>
-
-      
-      <div className={styles.introduce}>
-        <img src={myImage} width="300px" />
-        
-        <div className={styles.grid}>
-          <a href="https://nextjs.org/docs" className={styles.post}>
-            <p>先看看其他文章 &rarr;</p>
-          </a>
-        </div>
-      </div>
-     
-      <div className={styles.logincon}>
-        <p className={styles.emailpass}>使用者名稱：</p>
-          <input className={styles.enter} type="text" name="username" placeholder="(日後可更改)"/>
-        <p className={styles.emailpass}>常用信箱：</p>
-          <input className={styles.enter} type="text" name="email" value={account.email} onChange={handleChange} placeholder="請輸入信箱"/>
-        <p className={styles.emailpass}>密碼：</p>
-          <input className={styles.enter} type="password" name="password" value={account.password} onChange={handleChange} placeholder="請輸入密碼"/>
-        <p className={styles.emailpass}>確認密碼：</p>
-          <input className={styles.enter} type="password" name="password_check" placeholder="再次輸入密碼"/>
-        {message}
-        <input className={styles.loginbtn} type="button" onClick={haddleOnClick} value="註冊" name="register_btn"/>
-        <p>已經有帳號了嗎？</p>
-        <h4><a href="login.tsx">點此登入</a></h4>
-      </div>
-      
-      </main>
-   
-
-      <footer className={styles.footer}>
-        <a
-          href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Powered by{' '}
-          <span className={styles.logo}>
-            <Image src="/vercel.svg" alt="Vercel Logo" width={72} height={16} />
-          </span>
-        </a>
-      </footer>
-    </div>
-  )
-}
-export default Home;
-
-
