+++ styles/Home.module.css
@@ -708,7 +708,7 @@
   flex-direction: column;
 
   left:300px;
-  bottom:395px;
+  bottom:325px;
   height: 450px;
   width:500px;
   max-width: 400px;
